[Bronze III] Title: 직각삼각형, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ C++17/백준/Bronze/4153. 직각삼각형/직각삼각형.cc
@@ -2,7 +2,7 @@
 using namespace std;
 int main() {
 	while (1) {
-		long long int a, b, c, a2, b2, c2;
+		int a, b, c, a2, b2, c2;
 		cin >> a >> b >> c;
 		if (a == 0 && b == 0 && c == 0)
 			break;
